Location added
