+++ src/redux/constants/action.type.js
@@ -50,5 +50,6 @@ export const ActionTyps = {
     FETCH_VERIFIEDCOMPLAINTS_DASH: "FETCH_VERIFIEDCOMPLAINTS_DASH",
     FETCH_ONHOLDCOMPLAINTS_DASH: "FETCH_ONHOLDCOMPLAINTS_DASH",
     FETCH_EMP_DEPTSECTION_LIST: "FETCH_EMP_DEPTSECTION_LIST",
-    FETCH_TOTAL_ADMISSION: "FETCH_TOTAL_ADMISSION"
+    FETCH_TOTAL_ADMISSION: "FETCH_TOTAL_ADMISSION",
+    FETCH_EXTRAORDER_LIST: "FETCH_EXTRAORDER_LIST"
 }
