+++ src/views/Components/CusAgGridForMain.jsx
@@ -6,7 +6,7 @@ import 'ag-grid-community/dist/styles/ag-grid.css'
 import 'ag-grid-community/dist/styles/ag-theme-alpine.css'
 
 const CusAgGridForMain = ({ columnDefs, tableData, onSelectionChanged, columnTypes }) => {
-    const rowHeight = 30
+    const rowHeight = 35
     const headerHeight = 30
     const defaultColDef = {
     }
